fix(build/windows/macos): fix build caused by 40ca9ea80e
GitOrigin-RevId: b4868a08f63286f24f76ab457dbbc786a2e85936

+++ imperative/src/include/megbrain/imperative/profiler.h
@@ -40,8 +40,8 @@ struct ProfileEntry {
     std::vector<std::tuple<CompNode, TimeClosure, TimeClosure>> device_list;
     std::vector<ProfileTensor> inputs;
     std::vector<ProfileTensor> outputs;
-    ssize_t memory = 0;
-    ssize_t computation = 0;
+    long long memory = 0;
+    long long computation = 0;
 #if MGB_ENABLE_JSON
     std::shared_ptr<json::Value> param;
 #endif
